+++ contextlib2.py
@@ -4,7 +4,8 @@ import sys
 from collections import deque
 from functools import wraps
 
-__all__ = ["contextmanager", "closing", "ContextDecorator", "ContextStack"]
+__all__ = ["contextmanager", "closing", "ContextDecorator",
+           "ContextStack", "ExitStack"]
 
 
 class ContextDecorator(object):
@@ -142,12 +143,12 @@ class closing(object):
 
 
 # Inspired by discussions on http://bugs.python.org/issue13585
-class ContextStack(object):
+class ExitStack(object):
     """Context manager for dynamic management of a stack of exit callbacks
     
     For example:
     
-        with ContextStack() as stack:
+        with ExitStack() as stack:
             files = [stack.enter_context(open(fname)) for fname in filenames]
             # All opened files will automatically be closed at the end of
             # the with statement, even if attempts to open files later
@@ -155,23 +156,23 @@ class ContextStack(object):
     
     """
     def __init__(self):
-        self._callbacks = deque()
+        self._exit_callbacks = deque()
         
-    def preserve(self):
+    def pop_all(self):
         """Preserve the context stack by transferring it to a new instance"""
         new_stack = type(self)()
-        new_stack._callbacks = self._callbacks
-        self._callbacks = deque()
+        new_stack._exit_callbacks = self._exit_callbacks
+        self._exit_callbacks = deque()
         return new_stack
 
-    def _register_cm_exit(self, cm, cm_exit):
+    def _push_cm_exit(self, cm, cm_exit):
         """Helper to correctly register callbacks to __exit__ methods"""
         def _exit_wrapper(*exc_details):
             return cm_exit(cm, *exc_details)
         _exit_wrapper.__self__ = cm
-        self.register_exit(_exit_wrapper)
+        self.push(_exit_wrapper)
         
-    def register_exit(self, callback):
+    def push(self, exit):
         """Registers a callback with the standard __exit__ method signature
 
         Can suppress exceptions the same way __exit__ methods can.
@@ -179,16 +180,19 @@ class ContextStack(object):
         Also accepts any object with an __exit__ method (registering the
         method instead of the object itself)
         """
-        _cb_type = type(callback)
+        # We use an unbound method rather than a bound method to follow
+        # the standard lookup behaviour for special methods
+        _cb_type = type(exit)
         try:
-            exit = _cb_type.__exit__
+            exit_method = _cb_type.__exit__
         except AttributeError:
-            self._callbacks.append(callback)
+            # Not a context manager, so assume its a callable
+            self._exit_callbacks.append(exit)
         else:
-            self._register_cm_exit(callback, exit)
-        return callback # Allow use as a decorator
+            self._push_cm_exit(exit, exit_method)
+        return exit # Allow use as a decorator
 
-    def register(self, callback, *args, **kwds):
+    def callback(self, callback, *args, **kwds):
         """Registers an arbitrary callback and arguments.
         
         Cannot suppress exceptions.
@@ -198,19 +202,20 @@ class ContextStack(object):
         # We changed the signature, so using @wraps is not appropriate, but
         # setting __wrapped__ may still help with introspection
         _exit_wrapper.__wrapped__ = callback
-        self.register_exit(_exit_wrapper)
+        self.push(_exit_wrapper)
+        return callback # Allow use as a decorator
 
     def enter_context(self, cm):
         """Enters the supplied context manager
         
-        If successful, also registers its __exit__ method as a callback and
+        If successful, also pushes its __exit__ method as a callback and
         returns the result of the __enter__ method.
         """
         # We look up the special methods on the type to match the with statement
         _cm_type = type(cm)
         _exit = _cm_type.__exit__
         result = _cm_type.__enter__(cm)
-        self._register_cm_exit(cm, _exit)
+        self._push_cm_exit(cm, _exit)
         return result
 
     def close(self):
@@ -221,7 +226,7 @@ class ContextStack(object):
         return self
 
     def __exit__(self, *exc_details):
-        if not self._callbacks:
+        if not self._exit_callbacks:
             return
         # This looks complicated, but it is really just
         # setting up a chain of try-expect statements to ensure
@@ -230,8 +235,8 @@ class ContextStack(object):
         def _invoke_next_callback(exc_details):
             # Callbacks are removed from the list in FIFO order
             # but the recursion means they're invoked in LIFO order
-            cb = self._callbacks.popleft()
-            if not self._callbacks:
+            cb = self._exit_callbacks.popleft()
+            if not self._exit_callbacks:
                 # Innermost callback is invoked directly
                 return cb(*exc_details)
             # More callbacks left, so descend another level in the stack
@@ -250,3 +255,16 @@ class ContextStack(object):
             return suppress_exc
         # Kick off the recursive chain
         return _invoke_next_callback(exc_details)
+
+# Preserve backwards compatibility
+class ContextStack(ExitStack):
+    """Backwards compatibility alias for ExitStack"""
+
+    def register_exit(self, callback):
+        return self.push(callback)
+
+    def register(self, callback, *args, **kwds):
+        return self.callback(callback, *args, **kwds)
+
+    def preserve(self):
+        return self.pop_all()