+++ test_contextlib2.py
@@ -1,8 +1,11 @@
 #!/usr/bin/env python
-"""Unit tests for contextlib.py, and other context managers."""
+"""Unit tests for contextlib2"""
 
 import sys
+
 import unittest
+if not hasattr(unittest, "skipIf"):
+    import unittest2 as unittest
 
 from contextlib2 import *  # Tests __all__
 
@@ -300,6 +303,129 @@ class TestContextDecorator(unittest.TestCase):
         self.assertEqual(state, [1, 'something else', 999])
 
 
+class TestExitStack(unittest.TestCase):
+
+    def test_no_resources(self):
+        with ExitStack():
+            pass
+
+    def test_callback(self):
+        expected = [
+            ((), {}),
+            ((1,), {}),
+            ((1,2), {}),
+            ((), dict(example=1)),
+            ((1,), dict(example=1)),
+            ((1,2), dict(example=1)),
+        ]
+        result = []
+        def _exit(*args, **kwds):
+            """Test metadata propagation"""
+            result.append((args, kwds))
+        with ExitStack() as stack:
+            for args, kwds in reversed(expected):
+                if args and kwds:
+                    f = stack.callback(_exit, *args, **kwds)
+                elif args:
+                    f = stack.callback(_exit, *args)
+                elif kwds:
+                    f = stack.callback(_exit, **kwds)
+                else:
+                    f = stack.callback(_exit)
+                self.assertIs(f, _exit)
+            for wrapper in stack._exit_callbacks:
+                self.assertIs(wrapper.__wrapped__, _exit)
+                self.assertNotEqual(wrapper.__name__, _exit.__name__)
+                self.assertIsNone(wrapper.__doc__, _exit.__doc__)
+        self.assertEqual(result, expected)
+
+    def test_push(self):
+        exc_raised = ZeroDivisionError
+        def _expect_exc(exc_type, exc, exc_tb):
+            self.assertIs(exc_type, exc_raised)
+        def _suppress_exc(*exc_details):
+            return True
+        def _expect_ok(exc_type, exc, exc_tb):
+            self.assertIsNone(exc_type)
+            self.assertIsNone(exc)
+            self.assertIsNone(exc_tb)
+        class ExitCM(object):
+            def __init__(self, check_exc):
+                self.check_exc = check_exc
+            def __enter__(self):
+                self.fail("Should not be called!")
+            def __exit__(self, *exc_details):
+                self.check_exc(*exc_details)
+        with ExitStack() as stack:
+            stack.push(_expect_ok)
+            self.assertIs(stack._exit_callbacks[-1], _expect_ok)
+            cm = ExitCM(_expect_ok)
+            stack.push(cm)
+            self.assertIs(stack._exit_callbacks[-1].__self__, cm)
+            stack.push(_suppress_exc)
+            self.assertIs(stack._exit_callbacks[-1], _suppress_exc)
+            cm = ExitCM(_expect_exc)
+            stack.push(cm)
+            self.assertIs(stack._exit_callbacks[-1].__self__, cm)
+            stack.push(_expect_exc)
+            self.assertIs(stack._exit_callbacks[-1], _expect_exc)
+            stack.push(_expect_exc)
+            self.assertIs(stack._exit_callbacks[-1], _expect_exc)
+            1/0
+
+    def test_enter_context(self):
+        class TestCM(object):
+            def __enter__(self):
+                result.append(1)
+            def __exit__(self, *exc_details):
+                result.append(3)
+
+        result = []
+        cm = TestCM()
+        with ExitStack() as stack:
+            @stack.callback  # Registered first => cleaned up last
+            def _exit():
+                result.append(4)
+            self.assertIsNotNone(_exit)
+            stack.enter_context(cm)
+            self.assertIs(stack._exit_callbacks[-1].__self__, cm)
+            result.append(2)
+        self.assertEqual(result, [1, 2, 3, 4])
+
+    def test_close(self):
+        result = []
+        with ExitStack() as stack:
+            @stack.callback
+            def _exit():
+                result.append(1)
+            self.assertIsNotNone(_exit)
+            stack.close()
+            result.append(2)
+        self.assertEqual(result, [1, 2])
+
+    def test_pop_all(self):
+        result = []
+        with ExitStack() as stack:
+            @stack.callback
+            def _exit():
+                result.append(3)
+            self.assertIsNotNone(_exit)
+            new_stack = stack.pop_all()
+            result.append(1)
+        result.append(2)
+        new_stack.close()
+        self.assertEqual(result, [1, 2, 3])
+
+    def test_instance_bypass(self):
+        class Example(object): pass
+        cm = Example()
+        cm.__exit__ = object()
+        stack = ExitStack()
+        self.assertRaises(AttributeError, stack.enter_context, cm)
+        stack.push(cm)
+        self.assertIs(stack._exit_callbacks[-1], cm)
+
+
 class TestContextStack(unittest.TestCase):
     
     def test_no_resources(self):
@@ -322,14 +448,15 @@ class TestContextStack(unittest.TestCase):
         with ContextStack() as stack:
             for args, kwds in reversed(expected):
                 if args and kwds:
-                    self.assertIsNone(stack.register(_exit, *args, **kwds))
+                    f = stack.register(_exit, *args, **kwds)
                 elif args:
-                    self.assertIsNone(stack.register(_exit, *args))
+                    f = stack.register(_exit, *args)
                 elif kwds:
-                    self.assertIsNone(stack.register(_exit, **kwds))
+                    f = stack.register(_exit, **kwds)
                 else:
-                    self.assertIsNone(stack.register(_exit))
-            for wrapper in stack._callbacks:
+                    f = stack.register(_exit)
+                self.assertIs(f, _exit)
+            for wrapper in stack._exit_callbacks:
                 self.assertIs(wrapper.__wrapped__, _exit)
                 self.assertNotEqual(wrapper.__name__, _exit.__name__)
                 self.assertIsNone(wrapper.__doc__, _exit.__doc__)
@@ -354,19 +481,19 @@ class TestContextStack(unittest.TestCase):
                 self.check_exc(*exc_details)
         with ContextStack() as stack:
             stack.register_exit(_expect_ok)
-            self.assertIs(stack._callbacks[-1], _expect_ok)
+            self.assertIs(stack._exit_callbacks[-1], _expect_ok)
             cm = ExitCM(_expect_ok)
             stack.register_exit(cm)
-            self.assertIs(stack._callbacks[-1].__self__, cm)
+            self.assertIs(stack._exit_callbacks[-1].__self__, cm)
             stack.register_exit(_suppress_exc)
-            self.assertIs(stack._callbacks[-1], _suppress_exc)
+            self.assertIs(stack._exit_callbacks[-1], _suppress_exc)
             cm = ExitCM(_expect_exc)
             stack.register_exit(cm)
-            self.assertIs(stack._callbacks[-1].__self__, cm)
+            self.assertIs(stack._exit_callbacks[-1].__self__, cm)
             stack.register_exit(_expect_exc)
-            self.assertIs(stack._callbacks[-1], _expect_exc)
+            self.assertIs(stack._exit_callbacks[-1], _expect_exc)
             stack.register_exit(_expect_exc)
-            self.assertIs(stack._callbacks[-1], _expect_exc)
+            self.assertIs(stack._exit_callbacks[-1], _expect_exc)
             1/0
 
     def test_enter_context(self):
@@ -382,8 +509,9 @@ class TestContextStack(unittest.TestCase):
             @stack.register  # Registered first => cleaned up last
             def _exit():
                 result.append(4)
+            self.assertIsNotNone(_exit)
             stack.enter_context(cm)
-            self.assertIs(stack._callbacks[-1].__self__, cm)
+            self.assertIs(stack._exit_callbacks[-1].__self__, cm)
             result.append(2)
         self.assertEqual(result, [1, 2, 3, 4])
 
@@ -393,6 +521,7 @@ class TestContextStack(unittest.TestCase):
             @stack.register
             def _exit():
                 result.append(1)
+            self.assertIsNotNone(_exit)
             stack.close()
             result.append(2)
         self.assertEqual(result, [1, 2])
@@ -403,6 +532,7 @@ class TestContextStack(unittest.TestCase):
             @stack.register
             def _exit():
                 result.append(3)
+            self.assertIsNotNone(_exit)
             new_stack = stack.preserve()
             result.append(1)
         result.append(2)
@@ -416,7 +546,7 @@ class TestContextStack(unittest.TestCase):
         stack = ContextStack()
         self.assertRaises(AttributeError, stack.enter_context, cm)
         stack.register_exit(cm)
-        self.assertIs(stack._callbacks[-1], cm)
+        self.assertIs(stack._exit_callbacks[-1], cm)
         
 if __name__ == "__main__":
     import unittest