+++ NEWS.rst
@@ -1,10 +1,19 @@
 Release History
 ---------------
 
+0.4.0 (2012-05-05)
+~~~~~~~~~~~~~~~~~~
+
+* Issue #8: Replace ContextStack with ExitStack (old ContextStack API
+  retained for backwards compatibility)
+* Fall back to unittest2 if unittest is missing required functionality
+
+
 0.3.1 (2012-01-17)
 ~~~~~~~~~~~~~~~~~~
 
 * Issue #7: Add MANIFEST.in so PyPI package contains all relevant files
+
 
 0.3 (2012-01-04)
 ~~~~~~~~~~~~~~~~